requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --extra=dev --output-file=requirements-dev.txt
#
annotated-types==0.7.0
    # via pydantic
asgiref==3.8.1
    # via django
azure-core==1.30.2
    # via
    #   azure-storage-blob
    #   django-storages
azure-storage-blob==12.22.0
    # via django-storages
build==1.2.1
    # via pip-tools
certifi==2024.7.4
    # via requests
cffi==1.17.0
    # via cryptography
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   coreapi-cli
    #   pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
contourpy==1.3.0
    # via matplotlib
coreapi==2.3.3
    # via coreapi-cli
coreapi-cli==1.0.9
    # via Liionsden (pyproject.toml)
coreschema==0.0.4
    # via coreapi
crispy-bootstrap5==2024.2
    # via Liionsden (pyproject.toml)
cryptography==43.0.0
    # via
    #   Liionsden (pyproject.toml)
    #   azure-storage-blob
cycler==0.12.1
    # via matplotlib
debugpy==1.8.5
    # via Liionsden (pyproject.toml)
distlib==0.3.8
    # via virtualenv
django==5.1
    # via
    #   Liionsden (pyproject.toml)
    #   crispy-bootstrap5
    #   django-bootstrap-datepicker-plus
    #   django-bootstrap5
    #   django-crispy-forms
    #   django-extensions
    #   django-filter
    #   django-guardian
    #   django-js-asset
    #   django-storages
    #   django-tables2
    #   djangorestframework
    #   model-bakery
django-better-admin-arrayfield==1.4.2
    # via Liionsden (pyproject.toml)
django-bootstrap-datepicker-plus==5.0.5
    # via Liionsden (pyproject.toml)
django-bootstrap5==24.2
    # via Liionsden (pyproject.toml)
django-cleanup==8.1.0
    # via Liionsden (pyproject.toml)
django-crispy-forms==2.3
    # via
    #   Liionsden (pyproject.toml)
    #   crispy-bootstrap5
django-extensions==3.2.3
    # via Liionsden (pyproject.toml)
django-filter==24.3
    # via Liionsden (pyproject.toml)
django-guardian==2.4.0
    # via Liionsden (pyproject.toml)
django-ipware==7.0.1
    # via Liionsden (pyproject.toml)
django-js-asset==2.2.0
    # via django-mptt
django-jsoneditor==0.2.4
    # via Liionsden (pyproject.toml)
django-mptt==0.16.0
    # via Liionsden (pyproject.toml)
django-override-storage==0.3.2
    # via Liionsden (pyproject.toml)
django-storages[azure]==1.14.4
    # via Liionsden (pyproject.toml)
django-tables2==2.7.0
    # via Liionsden (pyproject.toml)
djangorestframework==3.15.2
    # via Liionsden (pyproject.toml)
et-xmlfile==1.1.0
    # via openpyxl
filelock==3.15.4
    # via virtualenv
fonttools==4.53.1
    # via matplotlib
identify==2.6.0
    # via pre-commit
idna==3.8
    # via requests
idutils==1.2.1
    # via Liionsden (pyproject.toml)
isbnlib==3.10.14
    # via idutils
isodate==0.6.1
    # via azure-storage-blob
itypes==1.2.0
    # via coreapi
jinja2==3.1.4
    # via coreschema
kiwisolver==1.4.5
    # via matplotlib
markupsafe==2.1.5
    # via jinja2
matplotlib==3.9.2
    # via preparenovonix
model-bakery==1.19.5
    # via Liionsden (pyproject.toml)
molmass==2024.5.24
    # via Liionsden (pyproject.toml)
mypy==1.11.2
    # via Liionsden (pyproject.toml)
mypy-extensions==1.0.0
    # via mypy
nodeenv==1.9.1
    # via pre-commit
numpy==2.1.0
    # via
    #   contourpy
    #   matplotlib
    #   pandas
    #   preparenovonix
oauthlib==3.2.2
    # via Liionsden (pyproject.toml)
openpyxl==3.1.5
    # via Liionsden (pyproject.toml)
packaging==24.1
    # via
    #   build
    #   django-jsoneditor
    #   matplotlib
    #   plotly
pandas==2.2.2
    # via Liionsden (pyproject.toml)
pillow==10.4.0
    # via
    #   Liionsden (pyproject.toml)
    #   matplotlib
pip-tools==7.4.1
    # via Liionsden (pyproject.toml)
platformdirs==4.2.2
    # via virtualenv
plotly==5.23.0
    # via Liionsden (pyproject.toml)
pre-commit==3.8.0
    # via Liionsden (pyproject.toml)
preparenovonix==1.0.3
    # via Liionsden (pyproject.toml)
psycopg2-binary==2.9.9
    # via Liionsden (pyproject.toml)
pycparser==2.22
    # via cffi
pydantic==2.8.2
    # via django-bootstrap-datepicker-plus
pydantic-core==2.20.1
    # via pydantic
pyparsing==3.1.4
    # via matplotlib
pyproject-hooks==1.1.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-ipware==3.0.0
    # via django-ipware
python-magic==0.4.27
    # via Liionsden (pyproject.toml)
pytz==2024.1
    # via pandas
pyyaml==6.0.2
    # via
    #   Liionsden (pyproject.toml)
    #   pre-commit
requests==2.32.3
    # via
    #   azure-core
    #   coreapi
ruff==0.6.2
    # via Liionsden (pyproject.toml)
six==1.16.0
    # via
    #   azure-core
    #   idutils
    #   isodate
    #   python-dateutil
sqlparse==0.5.1
    # via django
tablib==3.6.1
    # via Liionsden (pyproject.toml)
tenacity==9.0.0
    # via plotly
typing-extensions==4.12.2
    # via
    #   azure-core
    #   azure-storage-blob
    #   django-bootstrap-datepicker-plus
    #   mypy
    #   pydantic
    #   pydantic-core
tzdata==2024.1
    # via
    #   django
    #   pandas
uritemplate==4.1.1
    # via coreapi
urllib3==2.2.2
    # via requests
virtualenv==20.26.3
    # via pre-commit
wheel==0.44.0
    # via pip-tools
whitenoise==6.7.0
    # via Liionsden (pyproject.toml)
xlrd==2.0.1
    # via Liionsden (pyproject.toml)

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
